backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
Pillow==11.1.0
qrcode[pil]==8.0
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
django-storages==1.14.4
boto3==1.36.2
twilio==9.4.0
factory-boy==3.3.1
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
